Add files via upload
this is the code that is monitoring the components of computer system like keyboard, mouse, SSD etc. now help me to complete project that monitor logs of system if any component is changed or tempered. so it checks the previous logs and compare with newly generated logs. if any changes occur then it  sends alert to its system administrator